Webtoken

+++ src/main/webapp/controller/Users.java
@@ -15,8 +15,10 @@ import java.io.IOException;
 import java.io.PrintWriter;
 import java.security.NoSuchAlgorithmException;
 
+import static main.webapp.controller.Utils.isAuthorized;
+
 @WebServlet(name = "Home")
-public class Home extends HttpServlet {
+public class Users extends HttpServlet {
     private Gson gson = new Gson();
     private UserDao     userDao;
     public static final String CONF_DAO_FACTORY = "daofactory";
@@ -27,22 +29,31 @@ public class Home extends HttpServlet {
     }
 
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        String t = gson.toJson(userDao.findAll());
-        String mail = request.getParameter("mail");
         PrintWriter out = response.getWriter();
         response.setContentType("application/json");
         response.setCharacterEncoding("UTF-8");
-        if (mail != null ) {
-            out.print(gson.toJson(userDao.find(mail)));
+
+        if (isAuthorized(1, null, request, userDao)) {
+            String mail = request.getParameter("mail");
+
+            if (mail != null) {
+                out.print(gson.toJson(userDao.find(mail)));
+            } else {
+                out.print(gson.toJson(userDao.findAll()));
+            }
         }
         else {
-            out.print(gson.toJson(userDao.findAll()));
+            out.write("you're not allowed to do this, you must be admin");
         }
 
         out.flush();
     }
 
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        PrintWriter out = response.getWriter();
+        response.setContentType("application/json");
+        response.setCharacterEncoding("UTF-8");
+
         User toWrite = new User();
         String name = request.getParameter("name");
         String mail = request.getParameter("mail");
@@ -57,41 +68,52 @@ public class Home extends HttpServlet {
         toWrite.setPass(password);
         userDao.persist(toWrite);
 
-        String t = gson.toJson(toWrite);
-        PrintWriter out = response.getWriter();
-        response.setContentType("application/json");
-        response.setCharacterEncoding("UTF-8");
-        out.print(t);
+        out.print(gson.toJson(toWrite));
+
         out.flush();
     }
 
     @Override
     protected void doDelete(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        User toDelete = userDao.find(request.getParameter("mail"));
-        userDao.remove(toDelete);
-
-        String t = gson.toJson(toDelete);
         PrintWriter out = response.getWriter();
         response.setContentType("application/json");
         response.setCharacterEncoding("UTF-8");
-        out.print(t);
+
+        if (isAuthorized(1,null, request, userDao)) {
+            User toDelete = userDao.find(request.getParameter("mail"));
+            userDao.remove(toDelete);
+
+            out.print(gson.toJson(toDelete));
+        }
+        else {
+            out.write("you're not allowed to do this, you must be admin");
+        }
         out.flush();
     }
 
     @Override
     protected void doPut(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        User toUpdate = userDao.find(request.getParameter("mail"));
-        String password = request.getParameter("password");
-        String name = request.getParameter("name");
-        toUpdate.setPass(password);
-        toUpdate.setName(name);
-        userDao.update(toUpdate);
-
-        String t = gson.toJson(userDao.find(toUpdate.getMail()));
         PrintWriter out = response.getWriter();
         response.setContentType("application/json");
         response.setCharacterEncoding("UTF-8");
-        out.print(t);
+        User toUpdate = userDao.find(request.getParameter("mail"));
+        if (isAuthorized(2, toUpdate, request, userDao)) {
+            String password = null;
+            try {
+                password = Encoder.encode(request.getParameter("password"));
+            } catch (NoSuchAlgorithmException e) {
+                e.printStackTrace();
+            }
+            String name = request.getParameter("name");
+            toUpdate.setPass(password);
+            toUpdate.setName(name);
+            userDao.update(toUpdate);
+
+            out.print(gson.toJson(userDao.find(toUpdate.getMail())));
+        }
+        else {
+            out.write("you're not allowed to do this, you must be admin");
+        }
         out.flush();
     }
 }
